Clean some classes brought from the other project that semantically don't fit well in the jetpack clean architecture
