+++ composeApp/src/commonMain/kotlin/com.macaosoftware.component.navigationcompose.demo/di/DrawerViewModelsModule.kt
@@ -3,7 +3,6 @@ package com.macaosoftware.component.navigationcompose.demo.di
 import com.macaosoftware.component.core.RootDestinationRender
 import com.macaosoftware.component.drawer.DrawerComponentDefaults
 import com.macaosoftware.component.drawer.DrawerStatePresenterDefault
-import com.macaosoftware.component.navigationcompose.demo.marketplace.navigators.drawer.DemoDrawerDataSource
 import com.macaosoftware.component.navigationcompose.demo.marketplace.navigators.drawer.DemoDrawerRootDestinationRender
 import com.macaosoftware.component.navigationcompose.demo.marketplace.navigators.drawer.DemoDrawerViewModel
 import org.koin.compose.viewmodel.dsl.viewModelOf
@@ -13,7 +12,6 @@ import org.koin.dsl.module
 
 internal val drawerViewModelsModule = module {
 
-    factoryOf(::DemoDrawerDataSource)
     factory<DrawerStatePresenterDefault> {
         DrawerComponentDefaults.createDrawerStatePresenter()
     }
