+++ composeApp/src/commonMain/kotlin/com.macaosoftware.component.navigationcompose.demo/di/ServerUiModule.kt
@@ -2,16 +2,14 @@ package com.macaosoftware.component.navigationcompose.demo.di
 
 import com.macaosoftware.component.core.DestinationRendersRegistry
 import com.macaosoftware.component.navigationcompose.demo.marketplace.DemoDestinationRendersRegistry
-import com.macaosoftware.component.navigationcompose.demo.serverui.domain.ServerJsonManager
-import com.macaosoftware.component.navigationcompose.demo.serverui.domain.ServerUiDestinationInfoParser
-import org.koin.core.module.dsl.factoryOf
+import com.macaosoftware.component.navigationcompose.demo.serverui.domain.usecase.LoadChildrenDestinationUseCase
 import org.koin.dsl.module
 
 internal val serverUiModule = module {
 
+    factory<LoadChildrenDestinationUseCase> { LoadChildrenDestinationUseCase() }
+
     single<DestinationRendersRegistry> {
         DemoDestinationRendersRegistry()
     }
-    factoryOf(::ServerJsonManager)
-    factoryOf(::ServerUiDestinationInfoParser)
 }
